WIP: Create tabs and move volcano plot to `Volcano` tab
